fix function symbols in script.s
Otherwise the functions are not recognized as functions and the exported symbols are missing the 1 bit to mark them as thumb.
Happens to work out in the repo, because the functions are only ever called using bl, but the difference is notable when trying to use the symbols to do far calls using bx in romhacks.